backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.4,<5.0
